backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
greenlet==3.2.5
python-multipart==0.0.18
lxml==5.3.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
eval_type_backport==0.3.1

# GCP / production dependencies
asyncpg==0.30.0
psycopg2-binary==2.9.10
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.74.0
